Introduce "skippable" hard-quoted arguments
This is a first-cut implementation of a speculative feature for
letting a hard-quoted parameter be specified as skippable if the type
does not match.  It doesn't work on normal args:

    >> bad-skippy: func [x [<skip> integer!] y] [dump [x y]]
    ** Error: Only hard-quoted parameters are <skip>-able

But with a hard-quoted arg, it will be fulfilled if the type matches
or left as a null if not:

    >> skippy: func [:x [<skip> integer!] y] [dump [x y]]

    >> skippy "hi"
    x: => \\ null \\
    y: => "hi"

    >> skippy 10 "hi"
    x: => 10
    y: => "hi"

When enfixed, a skipped left argument will mean that a function will
not be executed greedily...it will run in its own step, as if the left
hand side was an end:

    >> lefty: enfix :skippy

    >> lefty "hi"
    x: => \\ null \\
    y: => "hi"

    >> evaluate [1 lefty "hi"]
    x: => 1
    y: => "hi"
    == []

    >> evaluate [<tag> lefty "hi"]
    == [lefty "hi"]

This also removes an error condition known as "lookback quote too late"
which would come up when a right-quoting operation outranked a left
quoting operation.  Instead, the left-quoted operation runs as a
separate operation, as if there's nothing on the left:

    >> quote 1 lefty "Hi"
    x: => \\ null \\
    y: => "Hi"

This gives a new opportunity to left-quoting operators to act like
non-enfix operations when they do not find what they want on their
left.  DEFAULT takes this opportunity, allowing it to just evaluate
to its right hand side whenever it doesn't see a SET-WORD! or a
SET-PATH! on its left, widening the scope of application.

    >> any [1 > 2 3 > 4 default ["hi"]]
    == "hi"

(Previously it would have complained that the quoted left argument it
found, 4, was not a SET-WORD! or a SET-PATH!.)