Cut down number of API entry points
This adds a native for constructing a DATE! out of year/month/day
and second/nano/timezone.  While somewhat inelegant, the Rebol native
avoids making and naming a C API which takes 6 C integer parameters--
instead having the date generation done by using rebI(...) to make
INTEGER! values to pass to the MAKE-DATE-YMDSNZ routine.  This is
a placeholder for a more elegant interface through MAKE DATE! which
was able to speak in terms of nanoseconds.

Similarly, it gets rid of routines specialized to the extraction of
nanoseconds from a date.  Instead, clients would use a Rebol routine
for getting an integer nanoseconds out of a time, and then use that.
(This is the general concept of the libRebol API...to raise the bar
on Rebol routines that should exist anyway.)

This eliminates the rebVersion() function, as it wasn't used, and the
information it reports could be extracted from the runtime via the
system object's version information.  Designing useful versioning on
top of the API would require something more sophisticated--and it's
a problem other APIs have (and usually do not solve).  C API versioning
is not something that Rebol has unique strengths to offer, so pursuing
invention in this area isn't a good investment of time.

rebIndexOf and rebLengthOf were hesitantly added, despite the fact that
`rebUnbox("index of", x, rebEND);` would work.  One reason they were
added was due to fears of user overrides breaking them, and these are
frequently used routines.  However, that problem is being solved by
being able to map natives to contexts to bind API calls from them from.
The idea that code would need to optimize INDEX OF and LENGTH OF but
not other things enough to use the internal API seems fairly unlikely,
so this goes ahead and made the (very few) calls use string calls.

The rebUnboxInteger/rebUnboxDecimal/rebUnboxChar functions did not use
a variadic interface, but this changes them so that they do...as well
as asserting the value is of the correct type.  This was made easier
by a recent change that factored out the variadic execution better.

With the advent of UTF-8 everywhere, conversions and aliases of TEXT!
to WORD! should be cheaper.  "as word!" and "as file!" should be able
to stand in acceptably, helping to keep the number of entry points
(especially ones for Wide Characters, which are only grudgingly
supported) to a minimum.

There was a locking API which was only used during boot, when the LOCK
native wasn't loaded yet.  It's not clear what advantage a C lock
API would have, but the version that was there was not useful...so this
gets rid of that for now and has boot use a static helper instead.