Drop Alloc on rebSpellAlloc/rebBytesAlloc, kill rebTimeNano
The JavaScript API wrapping dropped the Alloc() off of string routines
that allocated in the API, because they returned GC'd JavaScript items.
However, the motivation for putting the Alloc() on rebSpell() is to
remind users not to forget to free...but it's wordy, and you can still
forget.  Also, it may be that spellings are automatically GC'd for
you when the frame finishes.

In the interests of making the API more consistent and less wordy, this
drops the Alloc(), since the rest of the API doesn't put the note on
to remind you to rebRelease things.

Also, since MAKE-DATE-YMDSNZ exists, it solves the short-term need of
the ODBC module to create timestamps with nanoseconds.  The rebTimeNano
value constructor is thus not necessary, and removing it continues to
streamline the API's focus on minimal proxies from Rebol code to C.