Rework %make-reb-lib.r, allow no rebEND if C99/C++11 or later
This is a significant reworking of the %make-reb-lib.r script, which
was kind of redundant and brittle, and hard to change.  It had gone
through some improvement steps to use CSCAPE, but only incrementally.

As part of this process, it the EMIT-PROTOTYPES phase creates a list
of API representation objects.  Helpers like MAP-EACH-API allow for
code to generate items to be used in lists for each API routine.

Adds an enhancement to CSCAPE whereby BLANK! can be used to indicate a
substitution should be omitted, including that if blank substitutions
were the only content on a line then that line will be omitted.  It
doesn't work perfectly, but it helps make the code a bit clearer.

A big step here is that the API definitions are constructed in such a
way that if REBOL_IMPLICIT_END is defined before `#include "rebol.h"`
then it will make sure you are using C99 or C++11 (or later), and uses
the variadic macro feature to put rebEND onto the end of variadics
automatically.  Because this means all the rebXXX functions are now
macros, they can no longer have #ifdef's inside the calls, so there
had to be a fix to %mod-locale.c

This breaks the already-broken JavaScript CWRAP code a bit more, but
includes some basic changes just to show how the techniques like CWRAP
and MAP-EACH-API could be used there.