Make implicit ends default, reduce rebol.h dependencies
This takes some steps toward making it easier to include rebol.h from
external projects.  The goal is to make it possible for most users of
C99 or C++11 or higher to simply write something like this example that
uses Rebol to print out the numbered list of command line arguments to
a C program:

    #include "rebol.h"

    int main(int argc, char *argv[]) {
        rebStartup();

        int i;
        for (i = 0; i < argc; ++i)
           rebElide("print [{arg}", rebI(i), "is", rebT(argv[i]), "]");

        rebShutdown();
        return 0;
    }

What necessitates C99 is the design which allows the number of args
passed to a variadic like rebElide to be implicit.  This needs variadic
macros in C.  (It can be done in C++ without macros, though it's
somewhat more complex.)  Without that, one has to write something like:

  rebElide("print [{arg}", rebI(i), "is", rebT(argv[i]), "]", rebEND);

This commit changes the bias of the previous commit to default to
the implicit ends.  The core overrides this with REBOL_EXPLICIT_END
before `#include "rebol.h"`, so that when API calls are used in the
system they don't break buildability on older systems.

It shouldn't be necessary for clients of %rebol.h to use all the
prescriptive implementation choices of the core (including things like
defining `did` to `!!`, for instance).  In order to untangle that and
other core dependencies, this moves this reorders the library inclusion
to the top of %sys-core.h and %reb-host.h.

The reordering showed that rebol.h depends on some things that were in
%reb-c.h: attribute detection, noreturn/dead-end, GCC compiler version
checking convenience macros.  It also depends on RL_API which is a
macro that decorates the API prototypes for the RL_XXX functions
based on what kind of build is in effect.  These dependencies are
moved into %reb-config.h for the time being, though the next step will
be to review those and see what kind of minimal set might be put into
the rebol.h file itself to cover all the bases for non-core clients.

Some excessive comment separators are removed from %reb-c.h.