Use same opaque `struct Reb_Value` tag in C and C++
Early Ren-C implementations had some binary incompatibility between
cores built as C and C++, because the C++ build would have extra
checking bits that the C would corrupt (or not maintain).  All of those
checks have long since been accomplished other ways.

However, the exported name of the structure used in the API for
Reb_Value was different in C and in C++.  This was because the C++
classes added some derived behavior that wasn't necessarily desirable
in the base class.  Since C could only have one struct (no derivation)
and more than one struct would cause strict aliasing problems, this
meant that both C++'s Reb_Specific_Value and Reb_Relative_Value were
derived off of the class that served as the single Reb_Cell type in C.

Having a different struct tag name creates API complexity and chances
for the names used in rebol.h to be out of sync, if the flags that
control the choice are out of sync.  Picking which name to use based
on whether one is compiling as C++11 introduces a question of detection
and having that detection in the API match the detection used by
clients (or by the core).  This gets out of sync, for example, when
MSVC is used because they haven't bumped the __cplusplus standard
version...so the command line switches used in Rebol for CPLUSPLUS_11
overriding becomes needed by the API.

Commit also changes some long-winded explanations of what standard
layout structs are with a more effective assert that the structures for
Reb_Value and Reb_Relative value actually are standard layout, and
refers people to a link to read if those assertions happen to fail.