Eliminate union type punning in %t-vector.c, check range
R3-Alpha's VECTOR! datatype code has not been a focus or vetted very
much, but it had a number of sketchy things that it did.  One thing it
did was to use union type punning to create an "abstract API" over a
vector that could use a uintptr_t as a kind of universal currency
that could be either an integer of various sizes, or a decimal, or a
float:

https://en.wikipedia.org/wiki/Type_punning

Type punning is sketchy in C, and illegal in the C++ standard.  Plus
there is no real reason for it, vs. just using REBVAL as the currency
with either an INTEGER! or DECIMAL! value.

(It could be abstractly argued that float values may not be preserved
when round-tripping to decimal and back.  However, most floating point
values enter the system by way of Rebol DECIMAL! anyway.  The only way
this should be noticeable would be if someone loaded float data direct
into the system, ran something like a shuffle on it, and expected that
to not corrupt it...because reading a value out and writing it back
would potentially not preserve the exact bits.)

This goes ahead and switches the routines for getting and storing data
out of a vector to use REBVAL*.  It also avoids extracting the low
level vector properties (coming from a bitfield) to pass as parameters
and just lets the access routines read that information out of the
vector information.  This reduces the amount of code significantly.

Also, this adds bounds checking, so there will be a failure when
writing out-of-bounds data into an integer vector.