Cut down redundancy in symbol code, PROBE() fix
Investing an unrelated bug was made difficult by the amount of repeated
code in the symbol interning and table expansion...as well as the
use of the same term "hash" for both a linear probe calculated slot
and the original UTF-8 hash.  This distinguishes the terms with "slot"
and "hash", and factors repeated slot calculation out into an inline
function called First_Hash_Candidate_Slot().

Also, cuts down on some excessive commenting--which was useful at the
time of writing, but less so now that the code has hardened a bit.

The PROBE() code was testing for ARRAY_FLAG_VARLIST without first
making sure the series in question was an array--leading to debug
failures on symbol series, since the symbol number was interpreted
as series flags.  Puts varlist test after test for SERIES_FLAG_ARRAY.