Use full zero byte signal for END in values, endlike headers
A recent change allowed ends to be signaled by a clear bit in the
second byte.  This allowed for a short 1-character C string, whose
null terminator was in the second byte, to play the role of rebEND
in the API.

While a good change, it complicated matters since the bit was in the
inverse sense (a NOT_END bit).  Also, bits had to be masked out to
detect an end.  This raised the question of whether a zero byte could
always signal an END--for endlike headers and in value cells.

This commit does that by moving the series WIDE btye in the info from
the third byte to the second, and using a 0 byte to indicate that a
series is an array (real SER_WIDE() is given back as sizeof(REBVAL)).
It also moves the value's KIND byte (REB_BLOCK, REB_INTEGER, etc.)
into the second position, and takes REB_0 as a signal for END...moving
the previous usages out of band to REB_MAX + 1.