Use type punning in GCC little-endian builds for flags
The header bits are somewhat opaque, and the debugger is not always
cooperative when calling evaluations to try and test to see if a flag
is set with macros.  This changes `struct Reb_Header` into a union
which allows inspecting the header by ordered bytes in the debug build.

If using an appropriate GCC and endianness, it also defines packed
bitfield structures which allow looking into the header as series flags
or value flags or series ifo.