Eliminate SERIES_FLAG_ARRAY, use SER_WIDE() == 0
R3-Alpha originally did not have a separate flag for indicating arrays,
but recognized when SER_WIDE() of a series was equal to sizeof(REBVAL).
Auditing the system to get more checking on arrays was at a time of
expanding the number of bit flags, so it seemed easy enough to add
a distinct SERIES_FLAG_ARRAY...so that series could be used for non
arrays that happened to be the size of a a REBVAL on a 32 or 64 bit
platform.

With the change of a SER_WIDE() of 0 being used for arrays only (for
the purposes of planting a 0 byte in a position where it can serve as
an implicit array terminator), that is sufficient to distinguish an
array from a series whose item size just happens to be whatever a
REBVAL is.  So a bit can be reclaimed, and it may be on some platforms
checking for a zero byte is faster than masking the SERIES_FLAG_ARRAY
out of a platform-pointer sized integer.