Use small-series length byte as 255 to signal dynamic series
The small-series optimization allows a REBSER node to use the part of
the node where it would usually track a dynamic allocation, to store
the actual content itself--if it's small enough.  Making this more
useful is the ability to actually store a terminated Rebol array
in the size of one value cell, via the implicit termination trick
of SET_END.

This requires some signal in the series to indicate whether it is using
the small series optimization or not...otherwise SER_LEN() or
SER_DATA_AT_RAW() wouldn't know where to look.  Previously this was
done with a header bit of SERIES_FLAG_HAS_DYNAMIC.

Following in the pattern of the previous commit, this folds the flag
into a specialized state of the small series length byte...when it is
equal to 255.  The advantage of doing this is that in case the value
is not 255, there is no other memory location that needs to be looked
at to get the length...it can simply be returned.  This improves the
locality, and also reading/writing a byte may be faster than masking
bits on and off in some systems.

This couldn't "recover" a flag however, because there is still a reason
to keep a series flag related to dynamism...that is so that a call
to a series allocator can know when it is desired to *not* use the
small series optimization.  Things like paramlist arrays and object
contexts use this, because it's not terribly interesting to optimize
0-element contexts or 0-arg-or-local functions (the 1st element would
be consumed for the archetype cell, so it really would only be helping
zero cases.)

An added benefit to separating the flag from the state is that the
flag can indicate that a series should not be shrunk, even if there is
an opportunity for the GC to do so.  This couldn't be determined if the
flag for the allocation desire were unified with the state of whether
something held dynamic data or not.