REBCNT/REBINT use int_fast32_t
The community build of open-sourced R3-Alpha had a 64-bit conversion
(which has been progressively fixed for correctness in Ren-C).  However
that only fixed it so that 64-bits of address space could be used on
64-bit systems.  It did not convert the indexing of Rebol series to
use 64-bit counts on 64-bit systems.

Preparing the code to handle that would be somewhat involved, as
anywhere APIs are used that take smaller integers--as they relate to
Rebol indexing--could have a problem.  It's not terribly urgent to
do such a conversion.

However--an easier thing to do is to allow Rebol to keep using indexes
that aren't larger than 32-bits, but stored in the fastest variable
type for handling 32-bit numbers on that platform.  For some 64-bit
platforms that means using 64-bit numbers, as having values aligned
on 64-bit boundaries gives a boost.

This goes ahead and does that.  It required any code that was using
REBCNT as a name for a 32-bit unsigned int (or REBINT as a name for
a 32-bit signed int) to switch to uint32_t and int32_t (the types from
stdint.h) in order to keep working.  The cases where this was
a requirement were in things like file formats (BMP, GIF) and the
implementation code for the "deci" datatype (used by MONEY!)