Fix lingering REB_0_XXX names
Previously REB_0 was used as a kind of wildcard type for internal
algorithms that wanted to put custom payloads into cells.  However, it
turned out to be more useful for signifying ends of arrays.  So they
was moved to higher values past REB_MAX

This makes the relationship between these values and the Reb_Kind
enumeration a bit more clear, removing the REB_0 association.