Clean up strict aliasing workarounds and comments
The change of using byte-level access for IS_END() means that some
dicey workarounds to try and sync strict aliasing are no longer needed.

As a sign that getting rid of the workaround is probably a very good
idea, an attempt to remove the documentation of it to StackOverflow
seemed to cause a wave of confusion about what I was talking about:

But it was very real:

https://github.com/metaeducation/ren-c/issues/208
https://github.com/metaeducation/ren-c/issues/211

And it really did stop the problem:

https://github.com/metaeducation/ren-c/commit/f9014ce2298aed45a638fa89c2a86077e0c38b17

The new method is simpler, likely faster, and since unsigned char* has
no aliasing problems there should be no similar issue.