Use invisible mechanic for C-DEBUG-BREAK
C-DEBUG-BREAK was using a method of reevaluating which was hackier than
what can be done today simply using an invisible (e.g. return: [])
This makes the prototype more consistent with what a user would expect
they could do themselves with a similar operation.

Also, due to the working of invisibles, it was possible during an
argument evaluation to have a non-END argument in mid-fulfillment.  So
an optimization that was guided by the DO_FLAG_FULFILLING_ARG bit
(to avoid storing a temporary when evaluating GROUP!s) needed to be
changed to test if the output was END.