"psedutype" typeset flags, minor PARAM_CLASS speedup
Due to the exhaustion of flags for parameters, it was considered to
use extra bits in the typeset.  So for instance, since REB_0 wasn't
corresponding to an actual datatype, the bit was free to use for
endability instead of TYPESET_FLAG_ENDABLE.

This makes the change to expand the number of typeset flags (which
are available until REB_MAX hits 64).  It also tried to see if giving
the PARAM_CLASS the full byte instead of needing masking would make any
difference in performance, since VAL_PARAM_CLASS() is calculated often.

That didn't make much of a difference.  But still, the change gives
more options for distributing bits, and was somewhat clarifying vs the
macros like ALL_64, which were not compatible with this use of
the "pseudotypes" as flags.