Improve specialization method for Get_Var_Core()
The central routine for getting variables was parameterized with flags
which were tested to see if it needed to do extra checks.  However,
these specializations in -O2 optimization were not being factored out
to the point of removing those checks and flags on the calls that
didn't need them (e.g. read-only access).

This refactors it into a routine called Get_Context_Core(), which
lets the portion of the code that does mutability checking be elided
from read-only access routines.  This allows what was Get_Var_Core() to
be inlined fully into the evaluator.

Derelativize() is unchanged in this commit; merely moved to after
Get_Context_Core() so that they may wind up more easily sharing code.