Line up DO_FLAG_EVAL_FLIP with VALUE_FLAG_UNEVALUATED
This offers a bit of a performance improvement, by speeding a frequent
test of whether the "evaluating by default" state of the evaluator is
matching the evaluating state of individual values.  It does so with
about break-even clarity on the evaluator loop.

(The ability to tweak the "evaluating state" of values is reserved to
special modes...not conventional arrays.)

The use of XOR (^) instead of != appears to make a difference on 64-bit
architecture, at least in virtualbox.