Restore Get_Opt_Var_Or_End() vs. nullptr return
In performance analysis, it was noticed that Get_Opt_Var_Or_End()
didn't make much of a difference, and the confusing nature of
VAL_TYPE_OR_0 only applying on the global end marker was not worth
the potential confusion.

With the changing so that all end markers have a 0 byte in the VAL_TYPE
slot, it is no longer confusing.  The more confusing possibility is
actually that nullptr be confused for a NULL Rebol node.

Since it's really the same performance either way, this restores the
previous behavior, as it affords more "tricks" and disambiguates the
null case.