Pare complexity pertaining to frame end detection
The notion of frames having a "current value" underwent a change at
one time in which being at the end of a frame was distinguished from
having a frame's "current value" (f->value) be an END.  The original
change caused the introduction of a specialized class in the C++ build
to avoid the accidental appearance of END nodes in f->value.

It is now taken as a given that the IS_END() test can be collapsed with
a VAL_TYPE_RAW() test that returns 0, and that this will be efficient.
That makes it more appealing to focus on the fairly coherent idea that
when f->value is END, that is when a frame is at its end.