Expand f->eval_type to use pseudotypes vs. frame flags
A non-trivial amount of time in Eval_Core() on each entry is spent in
the first lines, where DO_FLAG_XXX are used to decide which point in
the code to `goto`.  This commit is based on the concept that if the
work involved in those branches could be folded in with other work that
had to be done anyway (e.g. making decisions for lookahead) then it
would reduce their overall cost.

This change is a first step in trying to cut down that branching.  The
f->eval_type is preloaded by the caller to Eval_Core(), and can be a
value beyond REB_MAX indicating the jump point.

There's no particular advantage to it in this change, but it tests the
concept.