Improve adherence to VALUE_FLAG_EVAL_FLIP
The EVAL_FLIP bit was introduced as an experiment, but later evolved
into something fundamental to the API model...to allow a mixture of
spliced "inert" values to combine with active ones.  It had never
really undergone a review of when it should be checked.

This makes the check itself more obvious by capturing it in a macro,
but also addresses missing cases of the check...as well as large
redundant code paths relating to SET-WORD! and SET-PATH!, which were
presumably added due to what were actually failures to handle the
flag fully on other evaluator branches.