Pass va_list by pointer-to-pointer in JavaScript wrapper
The argument vaptr is a *pointer* to a va_list, which needs to be
in some memory cell to be read.  Use a cell one past the END signal
in the argument list.

Also, changes rebEND to use a variable that currently leaks...but needs
to be allocated once in startup and then freed at shutdown.  This is
likely more efficient than having to construct the 2 byte sequence
manually each call.