Disallow evaluations with outstanding unmanaged objects
Originally in R3-Alpha, performing any evaluation would put series at
risk for GC--except for an arbitrary number of "orphans" which were
the most recent series created.  This unpredictable and error-prone
idea was replaced in Ren-C with managed and unmanaged series.  An
unmanaged series would not be GC'd, but rather would have to be
freed or promoted (or demoted?) to managed.  Once managed, a series
could only be freed by the garbage collector.

One issue however is what to do with the *content* of unmanaged arrays.
Though the array itself isn't a candidate for GC, its contents should
be kept live, or there isn't much of a purpose to the existence of
umanaged arrays...as you wind up having to guard them anyway.

However, some arrays are more than "just arrays"...they are self
referencing structures with complexities in their GC marking.  This
includes the varlist of contexts and the paramlist of actions.

Keeping these structures safe from GC with their arrays unmanaged is
difficult, if there's a goal that there be safety from referencing
unmanaged content from managed.  Since they are potentially deep
recursive structures with a lot of details, this commit moves to
instead mandate that no evaluations be performed while an unmanaged
context or paramlist is outstanding.  They have to be PUSH_GC_GUARD'd,
similar to what had to be done in R3-Alpha.