Reduce tax on GC for API "instructions"
As some lengthy comments said, the original desire was that when an
API instruction (such as rebEval()) was passed to a variadic, it would
automatically free the instruction as it went by.  This would mean the
GC would not be accumulating nodes, but quickly returning them back
into the seris pool.

However, the monitoring of manual series was in the way.  Refactoring
to easily allow the raw allocation of series nodes without lifetime
tracking was done for varlists of frames...and that could be reused
to update the API.