Simplify Drop_Frame() asserts, f->eval_type
Initially, the GC would react to a frame's eval_type to make finer
grained decisions about what in that frame to protect.  Problems with
errors arising at moments when something like a REB_ACTION had been
seen as an eval_type but weren't completely processed made using
the eval_type insufficient...and since frames were used for PARSE
it wasn't even the case that eval_type would be correct at all.

This means that there was a residual situation where REB_0 was being
used to mark frames that were "interstitial"...and had nothing
special to mark.  Saving and asserting that REB_0 condition added
computation, asserts, and explanation of the asserts.

This removes that, which makes things slightly faster and clearer.