Remove frame's eval_type field, revert to flags for gotos
In an attempt to reduce the overhead on every call into the evaluator,
a trick was tried to leverage the eval_type field, which already had
to be checked on each entry.  While a good idea in theory, it didn't
turn out to be any faster...possibly slower.

Instead, this commit restores the method using frame flag bits, and in
the continuing tidying/bugfixing, it removes the need for callers of
the evaluator to do any preparation of the eval_type.

Further: It actually *removes* the historical eval_type field from the
Reb_Frame structure completely.  That's because its original intent
(to allow the GC to selectively protect a frame based on what kind of
evaluation was taking place) no longer exists.  So all it was doing was
being an awkward cache which could get out of sync between parent and
child frames.

Instead, this uses an ordinary local which does not get mirrored from
parent frame to subframe...assuming that the cost of recalculating is
likely more than mirroring it into the subframe and then copying it
back.  That indeed does seem to be a bit faster--but more importantly,
it's less to get out of sync.