Remove eval_flag to simplify frame state, fix EVAL/ONLY
The `eval_flag` was a confusing concept, based on the idea that actions
could temporarily modify a frame's DO_FLAG_EXPLICIT_EVALUATE state.
With DO_FLAG_EVAL_FLIP being revisited and reviewed, this additional
complexity was determined to be solely used by the /ONLY refinement to
EVAL.

While the concept behind switching evaluation is now very heavily used
in the system and API, that particular ability for arbitrary functions
to change the containing frame's choices about evaluation has not seen
an application.  Additionally suggesting it might be a bad idea, the
implementation was buggy, with one of the three test cases actually
not behaving as it should (but having a bad test response filed so as
to go unnoticed).

In the push for continued simplification, this transforms the feature
into a generic ability to start subframes that jump to the reevaluation
point...which EVAL/ONLY uses.  The resulting code is easier to follow.