Revert to f->gotten as nullptr change
Enthusiasm for the new END abilities and concern over the confusion of
"nullptr" C with IS_NULLED() nodes led to experimenting with changing
the f->gotten from nullptr to END.

It's a tiny bit faster as nullptr, but that's not really the reason
to use it.  More relevant is that overuse of VAL_TYPE_RAW() means that
checking isn't just not done for END, but also not for unreadable
blanks.  When that's added in to the assignments being a bit slower,
it suggests changing it back.

The code is organized well enough now that this was fast to try and
fast to change back...and while doing so, to observe that code which
checks for both the enfixed bit and if something is an action is
redundant.  A small speedup comes from taking advantage of that.

Also includes updates to the comment explaining why enfixed is a
VALUE_FLAG_XXX and not an ACTION_FLAG_XXX.