Clean up GC interaction with deferred arguments
The introduction of deferred arguments meant that the evaluator could
be rewinding time, re-entering a subframe to give a second chance at
taking enfix processing.  However, the allowance of running Eval_Core()
into arguments that were previously considered "filled" allowed the
introduction of END cells as temporary states, confusing the GC which
thought they were already finalized.

An inefficient (and incorrect) way was used to deal with this by
changing the evaluator to back up its notion of enumeration state.
This commit updates the GC to understand the situation, allowing the
evaluator code to be more straightforward.

It also makes handling cells which may be ENDs more streamlined with
Queue_Mark_Opt_End_Cell_Deep, making some GC code dealing with values
easier to read.