Retake ACTION_FLAG_USER_NATIVE
There are only 8 bits available for ACTION_FLAGs, and two were taken
to indicate natives.  A need to reclaim one suggested folding the
detection of a user native in as fitting the pattern of a native, but
with a TEXT! in its source slot.

Browsing the file led to a slight bit of tinkering, to correct a bad
assertion and also to reduce the usage of internal data types (REBOOL,
REBCNT, etc.)  This is with an eye toward thinking about how the
balance of not using the internal API might be able to work for
something as low-level as TCC library interaction, but instead
using libRebol.