Turn f->doing_pickups into DO_FLAG_XXX
One longstanding thorn in the evaluator is that there has been no way
to fold the state of whether refinements are being "picked up" into
the f->refine field.  Hence it was an independent boolean.

Usage of the f->flags, a platform-sized integer used for control bits,
has grown over time for indicating such states.  It's better to use
that than a separate variable, considering it's already being consulted
so often that it is almost certainly in a register.

This has the advantage of making frames a little bit smaller and able
to be aligned better.