Eliminate path_type, use stored opt_setval to indicate SET
Frames previously held an "eval_type" that was phased out, with the
only remaining uses when a frame was holding state for path dispatch,
in order to know which kind of path it was (get-path!/path!/set-path!)

Since PATH! is just a null-intolerant form of GET-PATH!, it's no
longer needed for the dispatch process itself to know the difference
between the two.  And since f->special holds an optional value to
set to, that can just be checked for non-null to see if a path
dispatch is actually a set.