Simplify/accelerate SET-PATH! and SET-WORD!
When a SET-WORD! or SET-PATH! are running, there isn't very much frame
state in use.  So the concept of "Mid-Frame" evaluation was created to
be able to avoid the setup or teardown of a child frame, instead just
saving the frame's flag and data stack pointer...and doing the eval
of the right hand side into the same frame that the SET-WORD! or
SET-PATH! was running.

However, since the right hand side is evaluated for a SET-WORD! or
SET-PATH! before the left is used, there is a question about the
GC safety of the left-hand side.  If evaluating an array, one is given
safety due to the SERIES_INFO_HOLD plaed on it.  But if it's a va_list,
then the SET-WORD! or SET-PATH! might not have any references to it
besides the pointer that the evaluator had to it before running the
right hand side.

Erring on the side of caution, the evaluator would push the current
value to protect it.  This caution meant the frame's stack pointer had
to be saved and restored as well.  However, by changing the rule so
that all values passed to a variadic are responsible for their own
GC protection, this can be avoided...and the C stack level's `current`
serves as the "stack" for a chain of SET-WORD!s or SET-PATH!s, no
data stack needed.

Also adds an assert to Move_Value_Header() to make sure that a value
whose REB_XXX type is larger than REB_MAX (a "pseudotype") is never
moved...since there's no good reason to allow it, and it makes it
easier to debug rather than trying to figure out where the value came
from down the line.